Lesson learned:
Once you fully understand the problem, the solution is typically really short.
If you mostly understand the problem and then start trying to solve it,
your solution will typically be 2 or 3 times longer and more complicated than the
"optimal" solution.

Though that long solution may be write, that is not the one that an interviewer
wants